Adding thread pool to overlap faiss queries
